expression: fix negation
Basically the negation operator requires more external state tracking,
so store the last token in `token` use it to determine whether a '-' is
a negation operator or not.

Also fixes operator precedence when finishing a group by popping
negation off the stack if it is found after closing a group. It might
not be necessary, but it's fine since exponentiation isn't supported.